fitness tracker site/requirements.txt
autopep8==2.3.1
blinker==1.9.0
cachelib==0.13.0
click==8.1.7
colorama==0.4.6
Flask==3.1.0
Flask-Mail==0.10.0
Flask-Session==0.8.0
gunicorn==23.0.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
msgspec==0.18.6
packaging==24.2
pycodestyle==2.12.1
python-dotenv==1.0.1
vulture==2.14
Werkzeug==3.1.3
